Move FluxPointsEstimator
