+++ gammapy/estimators/__init__.py
@@ -5,3 +5,4 @@ from .lima_map import *
 from .ts_map import *
 from .asmooth_map import *
 from .sensitivity import *
+from .flux_point import *
