+++ gammapy/estimators/flux_point.py
